Refactored and linted
