+++ src/components/PageContent/index.tsx
@@ -1,0 +1,1 @@
+export { PageContent } from "./PageContent";
